refactor: move EIP-681 JNI exports into crate::jni
Introduce a `crate::jni` module tree to hold this library's `Java_*`
export surface, and move `RustEip681Tool`'s two exports into it as the
first submodule. `crate::eip681` keeps only the parsing logic and no
longer depends on the `jni` crate, so it is ordinary Rust that tests can
call directly.

Marshalling lives beside the exports it serves rather than in a shared
helpers module, so the `JniEip681TransactionRequest` encoders and
decoders move into `jni/eip681.rs` alongside the exports that use them.

Because the new module is named `jni`, a bare `jni::` path inside lib.rs
now resolves to it rather than to the `jni` crate. The 11 affected paths
there are qualified as `::jni::`. Submodules are unaffected, since a bare
path in a submodule already resolves to the crate.

Exported symbol names are unchanged, verified by diffing `nm -gU` output
against a pre-refactor baseline of 187 `Java_*` symbols. Kotlin binds to
these by name and resolves lazily, so a rename would be a runtime crash
rather than a compile error.
